requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==2.3.3
fsspec==2025.7.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
Pillow==10.0.1
postgrest==1.1.1
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==306
PyYAML==6.0.2
realtime==2.6.0
regex==2025.7.34
requests==2.31.0
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==13.1
Werkzeug==3.1.3
winshell==0.6
